+++ src/main/java/com/zerobase/bob/filesave/FileSave.java
@@ -1,4 +1,4 @@
-package com.zerobase.bob.review;
+package com.zerobase.bob.filesave;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
